% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenotypeParkinsonsDisease.R
\docType{package}
\name{PhenotypeParkinsonsDisease-package}
\alias{PhenotypeParkinsonsDisease}
\alias{PhenotypeParkinsonsDisease-package}
\title{PhenotypeParkinsonsDisease: A Study Package for Generating Cohort Diagnostics}
\description{
A package that generates cohort diagnostics for cohort definitions included in the package.
}
\keyword{internal}
